Enum xxcalc::linear_solver::SolvingError

pub enum SolvingError {
    NonLinear,
    NoSymbol,
    Tautology,
    NonSolvable,
}

An error that occurs during linear solving.

Variants
NonLinear

Provided expression is non linear (contains polynomials of degree greater than one).

NoSymbol

Expression contains only constants (no x symbol), so there is nothing to solve.

Tautology

The equation is solvable for any value of x (it is always true).

NonSolvable

The equation cannot be solved at all (as there is no such value of x which would make the equation true).