Struct xxcalc::parser::Parser

pub struct Parser { /* fields omitted */ }

Parser takes Tokens in infix form and transforms them into Reverse Polish Notation. After such transformation tokens can be processed by TokensReducer.

Parser stores registered operators (with their precedence and associativity) between multiple executions. Furthermore an output buffer of Tokens is kept internally, so it can be reused without allocating new memory from the operating system. If Parser lives long enough this behaviour can greatly reduce time wasted on mallocs.

Examples

let mut tokenizer = Tokenizer::default();
let mut parser = Parser::default();

parser.register_operator('+', Operator::new(1, OperatorAssociativity::Left));

let tokenized = tokenizer.process("2+2");
let parsed = parser.process(tokenized).unwrap();

assert_eq!(parsed.tokens, [(0, Token::Number(2.0)),
                           (2, Token::Number(2.0)),
                           (1, Token::Operator('+'))]);Run

Extending

One could embed the parser inside another TokensProcessor and initialize it there with some default operators.

Methods

impl Parser

fn new() -> Parser

Creates an empty Parser with no defined operators.

fn register_operator(&mut self, name: char, operator: Operator) -> Option<Operator>

Registers operator identified by given name to the parser.

Each operator must be registered and associated with its precedence and associative to make parser capable of valid RPN transformation. As operators always require two arguments, no arity is needed.

If an operator with given name was already registered, it previous specification is returned.

Examples

let mut tokenizer = Tokenizer::default();
let mut parser = Parser::default();

let r = parser.register_operator('+', Operator::new(42, OperatorAssociativity::Left));
assert!(r.is_none());

let r = parser.register_operator('+', Operator::new(1, OperatorAssociativity::Left));
assert_eq!(r.unwrap(), Operator::new(42, OperatorAssociativity::Left));
Run