Struct xxcalc::polynomial::Polynomial

pub struct Polynomial { /* fields omitted */ }

Polynomial struct represents a mathematical concept of polynomials.

A polynomial is defined by real values of its coefficients corresponding to natural-numbered powers of the variable x. As coefficients are represented as continous vector, storing polynomials of high degree, but with most terms empty is quite expensive (x^1023 makes vector of size 1024). A SmallVec is used, so coefficients up to power of 3 are stored on the stack.

As one can imagine x^3.14 is not a polynomial, while 3.14x^2+2 is a polynomial.

Examples

assert_eq!(String::from(Polynomial::new(&[1.0])), "1");
assert_eq!(String::from(Polynomial::new(&[0.0, 1.0])), "x");
assert_eq!(String::from(Polynomial::new(&[2.0, 0.0, 3.14])), "3.14x^2+2");Run

Methods

impl Polynomial

fn new(c: &[f64]) -> Polynomial

Creates new polynomial from given coefficients

Examples

assert_eq!(Polynomial::new(&[1.0, 2.0, 3.0])[0], 1.0);
assert_eq!(Polynomial::new(&[1.0, 2.0, 3.0])[2], 3.0);Run

fn linear(b: f64, a: f64) -> Polynomial

Creates a linear polynomial (a polynomial of degree one). Such polynomial represents a linear expression ax+b.

Examples

assert_eq!(Polynomial::linear(1.0, 0.0), Polynomial::new(&[1.0, 0.0]));
assert_eq!(Polynomial::linear(0.0, 1.0), Polynomial::new(&[0.0, 1.0]));Run

fn constant(b: f64) -> Polynomial

Creates a constant (degenerative) polynomial (a polynomial of degree zero).

Examples

assert_eq!(Polynomial::constant(0.0), Polynomial::new(&[0.0]));
assert_eq!(Polynomial::constant(3.14), Polynomial::new(&[3.14]));Run

fn zero() -> Polynomial

Creates a zero polynomial (a polynomial with constant 0.0). While the math does not define degree of such polynomial, for sake of usability, in this library such polynomial is defined with a degree zero.

Examples

assert_eq!(Polynomial::zero(), Polynomial::constant(0.0));
assert_eq!(Polynomial::zero().degree(), 0);Run

fn degree(&self) -> usize

Computes degree of the polynomial. This method has a linear complexity as degree is not stored precomputed.

Degree is a largest power with non-zero coefficient, however a zero polynomial has a zero degree. It should be observed that size of coefficients vector may not be equal to the degree, as some of coefficients may be zero.

Examples

assert_eq!(Polynomial::zero().degree(), 0);
assert_eq!(Polynomial::constant(0.0).degree(), 0);
assert_eq!(Polynomial::linear(1.0, 0.0).degree(), 0);
assert_eq!(Polynomial::linear(1.0, 1.0).degree(), 1);
assert_eq!(Polynomial::new(&[1.0, 1.0, 1.0, 1.0]).degree(), 3);
assert_eq!(Polynomial::new(&[1.0, 1.0, 1.0, 1.0, 0.0, 0.0]).degree(), 3);Run

fn bind(&self, x: f64) -> Polynomial

Computes value of the polynomial by binding the variable with given constant value.

A Horner method is used for this computation to keep it linear and efficient.

Examples

assert_eq!(Polynomial::constant(3.0).bind(9.0), Polynomial::constant(3.0));
assert_eq!(Polynomial::linear(0.0, 1.0).bind(2.0), Polynomial::constant(2.0));
assert_eq!(Polynomial::linear(1.0, 4.0).bind(2.0), Polynomial::constant(9.0));
assert_eq!(Polynomial::new(&[1.0, 4.0, 4.0]).bind(2.0), Polynomial::constant(25.0));Run

fn as_f64(&self) -> Result<f64, PolynomialError>

Tries to interpret polynomial as real value.

Polynomial can be transformed to real value (double) only if its degree is zero (if the polynomial is constant), otherwise value of polynomial can be binded for particular value of variable.

Errors

Will return PolynomialError::NonConstantError when the degree is greater than zero.

Examples

assert_eq!(Polynomial::constant(3.0).as_f64(), Ok(3.0));
assert_eq!(Polynomial::zero().as_f64(), Ok(0.0));
assert_eq!(Polynomial::linear(1.0, 0.0).as_f64(), Ok(1.0));
assert_eq!(Polynomial::linear(1.0, 1.0).as_f64(), Err(PolynomialError::NonConstantError));Run

fn as_string(&self, name: &str) -> String

Returns human readable representation of the polynomial. Uses provided name as name of the variable.

The returned string is as short as possible version of the polynomial starting with the largest exponents. Terms with empty coefficients are skipped, so as unnecessary symbols and exponents. In example 2x+1 is returned instead of 0*x^2+2*x^2+2*x^0.

Examples

assert_eq!(Polynomial::new(&[0.0, 2.0]).as_string("x"), "2x");
assert_eq!(Polynomial::new(&[-1.0, 2.0]).as_string("x"), "2x-1");
assert_eq!(Polynomial::new(&[-1.0, 2.0, 4.0]).as_string("x"), "4x^2+2x-1");
assert_eq!(Polynomial::new(&[-1.0, 2.0, 0.0, 4.0]).as_string("x"), "4x^3+2x-1");
assert_eq!(Polynomial::new(&[-1.0, 2.0, 0.0, 4.0]).as_string("foo"), "4foo^3+2foo-1");Run