Module xxcalc::polynomial_calculator::functions

Implementations of function handlers used in operators and functions.

This module is public, so when you are writing another evaluator, but it does not extend PolynomialEvaluator, you can reuse these functions as handlers for some of the common operations.

Operations that are directly defined on the Polynomial (such as addition, subtraction, multiplication and division), avoid cloning by using the assignment with operation, thus optimizing memory usage.

Functions

addition
Adds two polynomials together (requires two arguments).

bind
Binds a polynomial with a value (requires two arguments).

division
Divides two polynomials by each other (requires two arguments).

exponentiation
Performs exponentiation of polynomial (requires two arguments).

log
Computes a logarithm of arbitrary base (requires two constant arguments).

log10
Computes a decimal logarithm (requires a single constant argument).

multiplication
Multiplies two polynomials together (requires two arguments).

subtraction
Subtracts two polynomials from each other (requires two arguments).